+++ code/test/sixsq/nuvla/server/resources/module_application_lifecycle_test.clj
@@ -31,6 +31,8 @@
                   :author         "someone"
                   :commit         "wip"
 
+                  :unsupported-options  ["devices"]
+
                   :docker-compose "version: \"3.3\"\nservices:\n  web:\n    ..."})